pyproject.toml
[tool.poetry]
name = "ress_pomodoros"
version = "0.1.0"
description = ""
authors = ["Ress"]

[tool.poetry.dependencies]
python = "^3.9"
chalice = "^1.26.2"
pytest = "^6.2.5"
python-dotenv = "^0.19.2"
boto3 = "^1.20.13"
telebot = "^0.0.4"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
